+++ app/mainmenu.h
@@ -2,6 +2,7 @@
 #define MAINMENU_H
 
 #include "config/global.h"
+#include "config/local.h"
 
 #include <QObject>
 #include <QToolButton>
@@ -9,7 +10,7 @@
 class MainMenu : public QObject {
   Q_OBJECT
 public:
-  explicit MainMenu(QToolButton *btn, Config::Global &global_c);
+  explicit MainMenu(QToolButton *btn, Config::Global &global_c, Config::Local &local_c);
 
 signals:
   void exit();
@@ -24,6 +25,7 @@ private:
   QToolButton *button;
 
   Config::Global &global_conf;
+  Config::Local &local_conf;
 };
 
 #endif // MAINMENU_H